+++ src/core/imuPreIntegration.cpp
@@ -6,7 +6,7 @@ void ImuPreintegrator::Propagate(const Eigen::Vector3d ba, const Eigen::Vector3d
     {
         LOG(ERROR) << cv::format(
             "Invalid data for pre-integration propagate: start_ts_ = %f, end_ts_ = %f, imu_data_.size() = %d",
-            start_ts_, end_ts_, imu_data_.size());
+            start_ts_, end_ts_, static_cast<int>(imu_data_.size()));
         return;
     }
